% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddi.R
\name{get_latest_eddi_raster}
\alias{get_latest_eddi_raster}
\title{Download the Latest EDDI Raster from NOAA}
\usage{
get_latest_eddi_raster(timescale = "01wk", search_days = 14)
}
\arguments{
\item{timescale}{A character string for the desired timescale. Defaults to "01wk".}

\item{search_days}{An integer for how many past days to search for a valid file. Defaults to 14.}
}
\value{
A \code{SpatRaster} object from the \code{terra} package.
}
\description{
This function finds and downloads the most recently available EDDI (Evaporative
Demand Drought Index) raster file from the NOAA PSL data archive and loads it
as a SpatRaster object.
}
\examples{
\dontrun{
  # Get the latest 1-week EDDI raster
  latest_eddi_raster <- get_latest_eddi_raster()

  # Plot the raster
  terra::plot(latest_eddi_raster)
}
}
