% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ForDRI.R
\name{get_latest_fordri_raster}
\alias{get_latest_fordri_raster}
\title{Download the Latest ForDRI Raster from UNL}
\usage{
get_latest_fordri_raster(search_days = 14)
}
\arguments{
\item{search_days}{An integer for how many past days to search for a valid file. Defaults to 14.}
}
\value{
A \code{SpatRaster} object from the \code{terra} package.
}
\description{
This function finds and downloads the most recently available ForDRI (Flash
Drought Risk Index) raster file from the University of Nebraska-Lincoln's
data archive and loads it as a SpatRaster object.
}
\examples{
\dontrun{
  # Get the latest ForDRI raster
  latest_fordri_raster <- get_latest_fordri_raster()

  # Plot the raster
  terra::plot(latest_fordri_raster)
}
}
