% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather.R
\name{get_weather}
\alias{get_weather}
\title{Get Weather Observations}
\usage{
get_weather(station_ids, start_date, end_date, per_page = 1e+08, ...)
}
\arguments{
\item{station_ids}{A character vector of one or more station IDs.}

\item{start_date}{The start date for the query (can be a Date object or a string in "YYYY-MM-DD" format). The time will default to the beginning of the day (00:00:00).}

\item{end_date}{The end date for the query (can be a Date object or a string in "YYYY-MM-DD" format). The time will default to the end of the day (23:59:59).}

\item{per_page}{The number of records to return per page. Defaults to a very large number to retrieve all data in one request, mimicking the frontend.}

\item{...}{Additional GraphQL parameters (e.g., \code{page}, \code{sortBy}, \code{hasHistoricData}).}
}
\value{
A tidy tibble of weather observations.
}
\description{
Retrieves daily or hourly weather observations from the FEMS Climatology API
using the \code{weatherObs} GraphQL field.
}
\examples{
\dontrun{
  # Get weather data for a specific station
  wx_data <- get_weather(
    station_ids = "245410",
    start_date = "2024-07-01",
    end_date = "2024-07-05"
  )
}
}
