% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zentra.R
\name{get_zentracloud_v5_data_long}
\alias{get_zentracloud_v5_data_long}
\title{Get Long-Duration Timeseries Data from the ZentraCloud v5 API via Chunking}
\usage{
get_zentracloud_v5_data_long(
  device_id,
  start_datetime,
  end_datetime,
  token = get_zentracloud_apikey(),
  chunk_by = "7 days",
  parallel = FALSE
)
}
\arguments{
\item{device_id}{A character string for a single device serial number.}

\item{start_datetime}{The start date for the query.}

\item{end_datetime}{The end date for the query.}

\item{token}{Your ZentraCloud API token. Defaults to \code{get_zentracloud_apikey()}.}

\item{chunk_by}{A string for the time period of each chunk. Defaults to "7 days".}

\item{parallel}{A logical value. If TRUE, performs API calls in parallel.}
}
\value{
A tidy tibble of all successfully retrieved device readings.
}
\description{
Get Long-Duration Timeseries Data from the ZentraCloud v5 API via Chunking
}
