% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sofureport-package.R
\docType{package}
\name{sofureport-package}
\alias{sofureport}
\alias{sofureport-package}
\title{sofureport: Soil and Fuels Reporting}
\description{
Provides functions for pulling data from API's and creating reports.
}
\author{
\strong{Maintainer}: Josh Erickson \email{joshualerickson@gmail.com}

}
\keyword{internal}
