% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrangle_to_hourly}
\alias{wrangle_to_hourly}
\title{Floor to Hourly}
\usage{
wrangle_to_hourly(data, type = "synoptic")
}
\arguments{
\item{data}{A data.frame or data.table with 'date_time', 'station_id', and 'name' columns.}

\item{type}{A flag to include precip for 'zentra', default (NULL).}
}
\value{
A tibble that's been aggregated to an hourly summary using the mean. This works for precipitation because
synoptic pre computes accumulation in an hour with \code{precip_accum_one_hour_set_1} which is the same when taking the mean.
}
\description{
This function takes a data.frame, floors the 'date_time' to the hour,
and calculates the mean for all numeric columns by group.
}
