% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zentra.R
\name{zentra_format_datetime}
\alias{zentra_format_datetime}
\title{Get Timeseries Readings from the ZentraCloud v5 API}
\usage{
zentra_format_datetime(
  value,
  tz = Sys.getenv("ZENTRACLOUD_QUERY_TZ", "America/Denver")
)
}
\arguments{
\item{device_id}{A character string for a single device serial number.}

\item{start_datetime}{The start time for the query.}

\item{end_datetime}{The end time for the query.}

\item{token}{Your ZentraCloud API key. Defaults to \code{get_zentracloud_apikey()}.}

\item{units}{The units for the data. Defaults to "imperial".}

\item{...}{Additional query parameters to pass to the API.}
}
\value{
A tidy tibble of all readings.
}
\description{
Get Timeseries Readings from the ZentraCloud v5 API
}
